Took initial swing at connecting player and character tables up
