--- ExpressedRealms.DB/Character.cs
@@ -1,8 +1,0 @@
-﻿
-namespace ExpressedRealms.DB;
-
-public class Character
-{
-    public int Id { get; set; }
-    public string Name { get; set; } = null!;
-}
