+++ ExpressedRealms.DB/Characters/CharacterConfiguration.cs
@@ -1,0 +1,22 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace ExpressedRealms.DB.Characters;
+
+public class CharacterConfiguration : IEntityTypeConfiguration<Character>
+{
+    public void Configure(EntityTypeBuilder<Character> builder)
+    {
+        builder.HasKey(x => x.Id);
+        builder.Property(x => x.Id).IsRequired();
+        builder.Property(x => x.Name).IsRequired().HasMaxLength(150);
+        builder.Property(x => x.Background);
+        builder.Property(x => x.PlayerId).IsRequired();
+
+        builder.HasOne(x => x.Player)
+            .WithMany(x => x.Characters)
+            .HasForeignKey(x => x.PlayerId)
+            .OnDelete(DeleteBehavior.Restrict)
+            .IsRequired();
+    }
+}
